SourceCode/CipherSafeBackend/requirements.txt
uvicorn
fastapi
bcrypt
python-dotenv
SQLAlchemy
pydantic
python-jose
cryptography
psycopg2-binary
